feat : [서버] 마지막 로그인 기록 확인

+++ server/prisma/migrations/20251205182509_add_folder_sort_order/migration.sql
@@ -1,26 +1,21 @@
-/*
-  Warnings:
-
-  - You are about to drop the column `icon` on the `folders` table. All the data in the column will be lost.
-  - You are about to drop the `note_tags` table. If the table is not empty, all the data it contains will be lost.
-  - You are about to drop the `tags` table. If the table is not empty, all the data it contains will be lost.
-
-*/
--- DropForeignKey
-ALTER TABLE "public"."note_tags" DROP CONSTRAINT "note_tags_noteId_fkey";
-
--- DropForeignKey
-ALTER TABLE "public"."note_tags" DROP CONSTRAINT "note_tags_tagId_fkey";
-
--- DropForeignKey
-ALTER TABLE "public"."tags" DROP CONSTRAINT "tags_userId_fkey";
-
 -- AlterTable
-ALTER TABLE "folders" DROP COLUMN "icon",
-ADD COLUMN     "sortOrder" INTEGER;
+ALTER TABLE "folders" ADD COLUMN IF NOT EXISTS "sortOrder" INTEGER;
 
--- DropTable
-DROP TABLE "public"."note_tags";
+-- Drop icon column if exists
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT 1 FROM information_schema.columns
+        WHERE table_name = 'folders' AND column_name = 'icon'
+    ) THEN
+        ALTER TABLE "folders" DROP COLUMN "icon";
+    END IF;
+END $$;
 
--- DropTable
-DROP TABLE "public"."tags";
+-- Safely drop note_tags table if exists
+DROP TABLE IF EXISTS "public"."note_tags";
+DROP TABLE IF EXISTS "note_tags";
+
+-- Safely drop tags table if exists
+DROP TABLE IF EXISTS "public"."tags";
+DROP TABLE IF EXISTS "tags";
